GFPGAN/requirements.txt
basicsr>=1.4.2
facexlib>=0.2.5
realesrgan
lmdb
numpy
opencv-python
pyyaml
scipy
tb-nightly
torch>=1.7
torchvision
tqdm
yapf
uvicorn
fastapi
config
python-multipart
python-dotenv
minio
omegaconf
